Add the first test for 2D gaussian fitting
I took a snap-shot of the gui, and save it in a mat file. I also did a
quick fitting using a small set of AOIs and tracking their positions,
and the result is saved. The Fit_AOI callback is called by loading this
snap-shot, and the new output will be compared with the values saved
today.