Refactor: bypass the aoi boundary stuff, use coordinate
The original gaussian fitting function stores the boundary indices of
each AOI for slicing (xlow, xhigh, ylow, yhigh), making the code hard
to understand. I modified the getAOIsubImageAndCenterDuplicate to mimic
the AOI_Limit function, and directly return the subimage of the AOI.
The coordinates are read in each loop, either from the mapstruc_cell or
the data from the last frame.

Issues:
- getAOIsubImageAndCenterDuplicate assumes that the AOI center is not
  too closed to the edge of the image. Will raise index error if the AOI
  is too close (indexing out of the image).

- Bug found: The original logic didn't update the AOI coordinate between
  the first frame and the second frame, which means, in the second frame,
  the coordinate of the initial input AOI coordinate (those in mapstruc)
  is used to determine the AOI, rather than the fitted value from the
  first frame.

  I added a condition (framemapindx ~= 2) to preserve this bug,
  otherwise the test will break. Will remove that after I can make
  better tests.