Clean up slider1_callback
The CurrentField and indexing part does not seem necessary to me since
the slider number is rounded to integer at the beginning. The indexing
is somehow making sure that the increment is at least one, which will be
satisfied naturally by using rounded integers. I remove these code now
and see what happens.