Remove AOI mesh plot plotting feature
There was a cool feature to plot the AOI's intensity as a 3D mesh
surface. You can enter the AOI's center in the Xspot and Yspot textboxes,
and push the PlotContent button (shows "Image" text). After update graph
the gui will show the mesh plot, sized the aoiWidth, in axes1. Although
it is cool, manually entering the AOI's coordinates makes it hard to use.
I remove this feature for simplicity of the code base.